komplet fce vypoctu, ted uz "jen" vystup

+++ funkce.h
@@ -10,19 +10,19 @@ void        menu();
 char        odkudCtu();
 void        footer();
 void        velikostMatice(int &radku, int &sloupcu);
-void        makeMatrix(const int radku, const int sloupcu, std::vector<std::vector<long double> > &matice);
+void        makeMatrix(const int radku, const int sloupcu, std::vector<std::vector<double> > &matice);
 
 std::string filename();
 std::string overPriponu(std::string &jmeno);
 
 char        fileDelimiterFunction(char &delimiter);
-int         fileBranch(std::vector< std::vector <long double> > &matice);
-int         terminalFillMatrix(std::vector< std::vector <long double> > &matice);
-int         terminalBranch(std::vector< std::vector <long double> > &matice);
-void        printMatrix(std::vector< std::vector <long double> > &matice);
+int         fileBranch(std::vector< std::vector <double> > &matice);
+int         terminalFillMatrix(std::vector< std::vector <double> > &matice);
+int         terminalBranch(std::vector< std::vector <double> > &matice);
+void        printMatrix(std::vector< std::vector <double> > &matice);
 void        fileParseSizeOfMatrix(std::string &radek, int &errFlag, char &delimiter, int &rows, int &columns);
-int         fileFillLineOfMatrix(std::string &radek, std::vector<std::vector<long double> > &matice, const int i, char &delimiter);
-int         fileRead(std::string &filename, int &errFlag, char &delimiter, std::vector<std::vector<long double> > &matice);
+int         fileFillLineOfMatrix(std::string &radek, std::vector<std::vector<double> > &matice, const int i, char &delimiter);
+int         fileRead(std::string &filename, int &errFlag, char &delimiter, std::vector<std::vector<double> > &matice);
